feat!: initial code to start using cogs
